Web site/Flask/requirements.txt
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.3
distlib==0.3.3
filelock==3.3.1
Flask==2.0.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
platformdirs==2.4.0
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
virtualenv==20.9.0
Werkzeug==2.0.2
mysql-connector-python==8.0.31
Flask-WTF==1.0.1